backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pyodbc==5.0.1
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Image Processing (NEW for Phase 3)
Pillow==10.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0